deepen: fold onboarding-fallback notice into the Onboarding First Email module
Absorbs lib/email/onboarding-fallback-service.ts into
lib/onboarding/cached-summary-delivery.ts so the post-onboarding "first
filing email" is delivered through one interface.

Before this PR the [Onboarding] action ran a two-call protocol:
deliverFirstOnboardingEmail, then branch on reason ===
'no_cached_summaries' | 'no_tickers' and call sendOnboardingFallbackNotice.
The split exposed the fallback decision to caller branching that should
have been internal, and both modules wrote the same
User.onboardingFirstEmailSentAt idempotency column with no shared seam.

Dependency category: in-process (Prisma + email send). No port needed —
deepenable by merging the implementations and testing through the new
interface directly.

Tests deleted (12) at the absorbed seam:
- __tests__/lib/email/onboarding-fallback-service.test.ts

Tests added (8) at the deepened interface, in addition to keeping the
two existing pickBestSummaryForUser/cached-summary tests aligned with the
new no-cache → fallback flow:
- __tests__/lib/onboarding/cached-summary-delivery.test.ts new
  "fallback path (no cached summary)" describe block covering subject-line
  truncation (≤3 / >3 / 0 tickers), the type:onboarding-fallback-notice
  tag, idempotency persistence, fallback_failed on sendEmail.success=false,
  and internal_error on getEmailTemplate / sendEmail throw.

CONTEXT.md adds an "Onboarding First Email" entry naming the deepened
module, the absorbed two-call protocol, and the DeliveryResult.reason
discriminator. Same shape as Template Selection / Filing Content Verifier
/ SEC Section — one deep module behind one function, form-type-style
branching kept private.
