+++ CONTEXT.md
@@ -24,6 +24,9 @@ A user's saved interest in a [Filing] type or ticker. Drives which [Summary] pay
 ### Onboarding
 The first-run flow that turns a new signup into an engaged subscriber: ticker selection, preview email, account creation, first cached [Summary] delivery.
 
+### Onboarding First Email
+The post-onboarding "first filing email" delivered through Next.js `after()` once the signup action returns. Implemented as one deep module (`lib/onboarding/cached-summary-delivery.ts`) behind a single `deliverFirstOnboardingEmail(userId, userEmail, recipientName)` interface that owns BOTH delivery paths: the cached-summary happy path (rank the user's tickers' cached [Summary] entries, pick the best, send via `sendFilingSummaryEmail`) AND the "we're watching your tickers" fallback that runs when no cached candidate exists. Idempotent on `User.onboardingFirstEmailSentAt` — whichever path delivers writes the same column, so re-running for the same user is always a no-op regardless of which path ran. Returns one `DeliveryResult` whose `reason` discriminates outcomes (`already_sent`, `email_failed`, `fallback_sent`, `fallback_failed`, `internal_error`). Replaces a former two-call protocol where the [Onboarding] action had to call `deliverFirstOnboardingEmail` first, branch on `reason === 'no_cached_summaries' | 'no_tickers'`, and call a separate `sendOnboardingFallbackNotice` from `lib/email/onboarding-fallback-service.ts` — the split forced every caller to remember the second send and exposed the fallback decision to caller branching that should have been internal. The fallback's subject-line truncation (first 3 tickers + "+N" overflow, generic copy when zero tickers) and template tag (`type:onboarding-fallback-notice`) are private to the module; the [Filing Summary Email] tag (`type:filing-summary` family) is owned by `sendFilingSummaryEmail`. Tests at `__tests__/lib/onboarding/cached-summary-delivery.test.ts` exercise both paths through the same interface.
+
 ### X Sentiment
 An enrichment of a [Summary] that captures public discussion on X (Twitter) about the ticker over a 7-day window. Direction (bullish/bearish/mixed/neutral/no_signal), shift, confidence, and citation-backed claims. Gated by form importance + a large-cap allowlist (S&P 500) to keep cost and pump-and-dump risk bounded. Runs only for paid-tier callers via the enrichment gate.
